Fix compilation error

+++ src/CScriptParams.cpp
@@ -71,14 +71,14 @@ void CScriptParams::DetectError()
 	
 	if (m_flags & Flags::MORE_PARAMETER_ALLOWED)
 	{
-		if (m_params[0] < (m_paramscount * sizeof(cell)))
+		if (m_params[0] < (m_paramscount * static_cast<int>(sizeof(cell))))
 		{
 			m_error = PARAMETER_COUNT_ISNT_ENOUGH;
 		}
 	}
 	else
 	{
-		if (m_params[0] != (m_paramscount * sizeof(cell)))
+		if (m_params[0] != (m_paramscount * static_cast<int>(sizeof(cell))))
 		{
 			m_error = PARAMETER_COUNT_ISNT_EQUAL;
 		}
